# openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequestDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/LoginResponseDto" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Answer": {
        "required": ["content", "isCorrect", "questionId"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "content": { "maxLength": 5000, "minLength": 5, "type": "string" },
          "isCorrect": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "questionId": { "type": "string", "format": "uuid" },
          "question": { "$ref": "#/components/schemas/Question" }
        }
      },
      "Question": {
        "required": ["content", "questionType"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "content": { "maxLength": 5000, "minLength": 5, "type": "string" },
          "questionType": {
            "type": "string",
            "enum": [
              "MultipleChoice",
              "SingleChoice",
              "TrueFalse",
              "FillInTheBlanks",
              "ShortAnswer",
              "LongAnswer"
            ]
          },
          "isActive": { "type": "boolean" },
          "quizQuestions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/QuizQuestion" }
          },
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Answer" }
          }
        }
      },
      "Quiz": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "duration": { "type": "integer", "format": "int32" },
          "thumbnailUrl": { "type": "string" },
          "isActive": { "type": "boolean" },
          "userQuizzes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserQuiz" }
          },
          "quizQuestions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/QuizQuestion" }
          }
        }
      },
      "QuizQuestion": {
        "required": ["questionId", "quizId"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "quizId": { "type": "string", "format": "uuid" },
          "quiz": { "$ref": "#/components/schemas/Quiz" },
          "questionId": { "type": "string", "format": "uuid" },
          "question": { "$ref": "#/components/schemas/Question" },
          "order": { "type": "integer", "format": "int32" }
        }
      },
      "Role": {
        "required": ["description", "name"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "maxLength": 50, "minLength": 3, "type": "string" },
          "description": { "maxLength": 50, "minLength": 3, "type": "string" },
          "isActive": { "type": "boolean" }
        }
      },
      "User": {
        "required": ["email", "firstName", "lastName", "password"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "firstName": { "maxLength": 50, "minLength": 3, "type": "string" },
          "lastName": { "maxLength": 50, "minLength": 3, "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "avatar": { "maxLength": 500, "minLength": 0, "type": "string" },
          "dateOfBirth": { "type": "string", "format": "date" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "userQuizzes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserQuiz" }
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": { "$ref": "#/components/schemas/Role" }
          },
          "displayName": { "type": "string" }
        }
      },
      "UserAnswer": {
        "required": ["answerId", "questionId", "userQuizId"],
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userQuizId": { "type": "string", "format": "uuid" },
          "userQuiz": { "$ref": "#/components/schemas/UserQuiz" },
          "questionId": { "type": "string", "format": "uuid" },
          "question": { "$ref": "#/components/schemas/Question" },
          "answerId": { "type": "string", "format": "uuid" },
          "answer": { "$ref": "#/components/schemas/Answer" },
          "isCorrect": { "type": "boolean" }
        }
      },
      "UserQuiz": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "user": { "$ref": "#/components/schemas/User" },
          "quizId": { "type": "string", "format": "uuid" },
          "quiz": { "$ref": "#/components/schemas/Quiz" },
          "quizCode": { "type": "string" },
          "startedAt": { "type": "string", "format": "date-time" },
          "finishedAt": { "type": "string", "format": "date-time" },
          "userAnswers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserAnswer" }
          }
        }
      },
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" },
          "fullName": { "type": "string" }
        }
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "LoginResponseDto": {
        "required": ["accessToken", "refreshToken"],
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "$ref": "#/components/schemas/RefreshToken" }
        }
      },
      "RefreshToken": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "expiryDate": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}